Refactor export functions to static methods
