+++ libraries/classes/Plugins/Export/ExportPdf.php
@@ -8,6 +8,7 @@
  */
 namespace PhpMyAdmin\Plugins\Export;
 
+use PhpMyAdmin\Export;
 use PhpMyAdmin\Plugins\ExportPlugin;
 use PhpMyAdmin\Plugins\Export\Helpers\Pdf;
 use PhpMyAdmin\Properties\Plugins\ExportPluginProperties;
@@ -150,7 +151,7 @@ class ExportPdf extends ExportPlugin
         $pdf = $this->_getPdf();
 
         // instead of $pdf->Output():
-        if (!PMA_exportOutputHandler($pdf->getPDFData())) {
+        if (!Export::outputHandler($pdf->getPDFData())) {
             return false;
         }
