+++ test/classes/plugin/export/ExportSqlTest.php
@@ -8,7 +8,6 @@
 use PhpMyAdmin\Plugins\Export\ExportSql;
 use PhpMyAdmin\Table;
 
-require_once 'libraries/export.lib.php';
 require_once 'libraries/config.default.php';
 require_once 'test/PMATestCase.php';
