+++ test/classes/plugin/export/ExportYamlTest.php
@@ -7,7 +7,6 @@
  */
 use PhpMyAdmin\Plugins\Export\ExportYaml;
 
-require_once 'libraries/export.lib.php';
 require_once 'libraries/config.default.php';
 require_once 'libraries/database_interface.inc.php';
 require_once 'test/PMATestCase.php';
